feat: 주문 상품 데이터 패칭 로직 작성 및 리팩토링
