+++ src/components/OrderTable/OrderTable.module.scss
@@ -1,0 +1,24 @@
+.table-container {
+  max-height: 500px;
+  overflow-y: auto;
+  width: 100%;
+  margin: 0 auto 20px;
+  display: flex;
+  flex-direction: column;
+
+  tr {
+    width: 100%;
+    border: 10px solid transparent;
+  }
+
+  tbody {
+    overflow-y: auto;
+  }
+}
+
+.table-header {
+  th {
+    border-bottom: 1px solid $yellow;
+    color: $dark-gray;
+  }
+}
