+++ src/containers/Order/Order/Order.module.scss
@@ -17,6 +17,12 @@
   margin: 0 auto 50px;
   display: flex;
   flex-direction: column;
+
+  .total-price {
+    display: flex;
+    justify-content: flex-end;
+    margin-bottom: 20px;
+  }
 }
 
 .info-layout {
